requirements.txt
langchain
langchain-community
langchain-core
langgraph
langchain-text-splitters
langchain-groq
langchain-google-genai
chromadb
pypdf
python-dotenv